requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dhooks==1.1.4
discord-protos==0.0.2
discord.py-self==2.0.0
frozenlist==1.4.1
idna==3.10
joblib==1.4.2
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.2
regex==2024.9.11
requests==2.32.3
tqdm==4.66.5
urllib3==2.2.3
yarl==1.14.0
